SOLR-5986: Add an ExitableDirectoryReader in Lucene and use that to support exiting of long running queries in Solr. (Merge from trunk r1627622)
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/branch_5x@1627635 13f79535-47bb-0310-9956-ffa450edef68

+++ solr/core/src/java/org/apache/solr/handler/component/QueryComponent.java
@@ -285,7 +285,7 @@ public class QueryComponent extends SearchComponent
     }
 
     // -1 as flag if not set.
-    long timeAllowed = (long)params.getInt( CommonParams.TIME_ALLOWED, -1 );
+    long timeAllowed = params.getLong(CommonParams.TIME_ALLOWED, -1L);
     if (null != rb.getCursorMark() && 0 < timeAllowed) {
       // fundementally incompatible
       throw new SolrException(SolrException.ErrorCode.BAD_REQUEST, "Can not search using both " +
